part 1 : added a line to clear the player character after drawing it
The line is removed in the last code sample but is never added prior to that.

It might make more sense to have a section where the player is able to move around but the line is absent and they leave behind a trail.
This would demonstrate why a space is put in the player's location after rendering and provide an opportunity to mention methods of clearing the console.